+++ src/main/gen/icu/windea/pls/localisation/psi/impl/ParadoxLocalisationStellarisNamePartImpl.java
@@ -1,15 +1,12 @@
 // This is a generated file. Not intended for manual editing.
 package icu.windea.pls.localisation.psi.impl;
 
-import java.util.List;
 import org.jetbrains.annotations.*;
 import com.intellij.lang.ASTNode;
 import com.intellij.psi.PsiElement;
 import com.intellij.psi.PsiElementVisitor;
-import com.intellij.psi.util.PsiTreeUtil;
-import static icu.windea.pls.localisation.psi.ParadoxLocalisationElementTypes.*;
 import icu.windea.pls.localisation.psi.*;
-import icu.windea.pls.localisation.references.ParadoxLocalisationStellarisNameFormatPsiReference;
+import icu.windea.pls.localisation.references.ParadoxLocalisationStellarisNamePartPsiReference;
 
 public class ParadoxLocalisationStellarisNamePartImpl extends ParadoxLocalisationRichTextImpl implements ParadoxLocalisationStellarisNamePart {
 
@@ -42,7 +39,7 @@ public class ParadoxLocalisationStellarisNamePartImpl extends ParadoxLocalisatio
 
   @Override
   @Nullable
-  public ParadoxLocalisationStellarisNameFormatPsiReference getReference() {
+  public ParadoxLocalisationStellarisNamePartPsiReference getReference() {
     return ParadoxLocalisationPsiImplUtil.getReference(this);
   }